[chore] Extension 폴더 분리 #9
