--- FindTown/FindTownUI/Sources/FindTownUI/UIColor+Extension.swift
@@ -1,34 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by 장선영 on 2022/12/16.
-//
-
-import Foundation
-import UIKit
-
-public extension UIColor {
-    
-    static let back1: UIColor = .init(named: "Back1", in: .module, compatibleWith: nil) ?? .clear
-    static let back2: UIColor = .init(named: "Back2", in: .module, compatibleWith: nil) ?? .clear
-    static let back3: UIColor = .init(named: "Back3", in: .module, compatibleWith: nil) ?? .clear
-    
-    static let black: UIColor = .init(named: "Black", in: .module, compatibleWith: nil) ?? .clear
-    static let dim20: UIColor = .init(named: "Dim20", in: .module, compatibleWith: nil) ?? .clear
-    static let dim50: UIColor = .init(named: "Dim50", in: .module, compatibleWith: nil) ?? .clear
-    static let dim70: UIColor = .init(named: "Dim70", in: .module, compatibleWith: nil) ?? .clear
-    
-    static let grey1: UIColor = .init(named: "Grey1", in: .module, compatibleWith: nil) ?? .clear
-    static let grey2: UIColor = .init(named: "Grey2", in: .module, compatibleWith: nil) ?? .clear
-    static let grey3: UIColor = .init(named: "Grey3", in: .module, compatibleWith: nil) ?? .clear
-    static let grey4: UIColor = .init(named: "Grey4", in: .module, compatibleWith: nil) ?? .clear
-    static let grey5: UIColor = .init(named: "Grey5", in: .module, compatibleWith: nil) ?? .clear
-    static let grey6: UIColor = .init(named: "Grey6", in: .module, compatibleWith: nil) ?? .clear
-    static let grey7: UIColor = .init(named: "Grey7", in: .module, compatibleWith: nil) ?? .clear
-    
-    static let primary: UIColor = .init(named: "Primary", in: .module, compatibleWith: nil) ?? .clear
-    static let secondary: UIColor = .init(named: "Secondary", in: .module, compatibleWith: nil) ?? .clear
-    static let white: UIColor = .init(named: "White", in: .module, compatibleWith: nil) ?? .clear
-    
-}
